Remove console.log
