+++ twake/backend/node/src/core/platform/services/database/services/orm/connectors/mongodb/query-builder.ts
@@ -1,6 +1,6 @@
 import { FindOptions } from "../../repository/repository";
 import { ObjectType } from "../../types";
-import { getEntityDefinition } from "../../utils";
+import { getEntityDefinition, secureOperators } from "../../utils";
 import { transformValueToDbString } from "./typeTransforms";
 
 export function buildSelectQuery<Entity>(
@@ -34,48 +34,11 @@ export function buildSelectQuery<Entity>(
     }
   });
 
-  findOptions = secureOperators(findOptions, entityType, options);
+  findOptions = secureOperators(transformValueToDbString, findOptions, entityType, options);
   where = buildComparison(where, findOptions);
   where = buildIn(where, findOptions);
 
   return where;
-}
-
-export function secureOperators<Entity>(
-  findOptions: FindOptions = {},
-  entityType: ObjectType<Entity>,
-  options: {
-    secret?: string;
-    keyspace: string;
-  } = {
-    secret: "",
-    keyspace: "twake",
-  },
-): FindOptions {
-  const instance = new (entityType as any)();
-  const { columnsDefinition, entityDefinition } = getEntityDefinition(instance);
-
-  Object.keys(findOptions).forEach(key => {
-    if (
-      key == "$in" ||
-      key == "$lte" ||
-      key == "$lt" ||
-      key == "$gte" ||
-      key == "$gt" ||
-      key == "$like"
-    ) {
-      findOptions[key].forEach(element => {
-        element[1] = element[1].map((e: any) =>
-          transformValueToDbString(e, columnsDefinition[element[0]].type, {
-            columns: columnsDefinition[element[0]].options,
-            secret: options.secret || "",
-          }),
-        );
-      });
-    }
-  });
-
-  return findOptions;
 }
 
 export function buildComparison(where: any, options: FindOptions = {}): string[] {
